JFrame Setup
The "CookManager" class has been setup such that an empty window is created by the program when it is run. Panel classes have been added to later be built off. Images that may be used later have been added into a "res folder". A "CookSettings" file has been added to provide a central area from which to access standard variables. Other residual files have been added to the program for the construction stages