List Saving and Deletion
CookFood is now able to save a list of given ingredients to a user-input location on the computer. The structure of the enum "QuantityType", was altered to reflect this, and allow for the output of a list that is syntactically consistent with the English language. The ability to delete a recipe from system storage given it's name was also added