Buttons and Icons
Did some work on adding four buttons to the main screen. They currently appear with text that scales based on the size of the frame. The size of the buttons and the whitespace around them also now scales with the frame as it is resizes. More work needs to be done on the "CookButton" class to ensure that JButtons can be added to the frame that have icons that are more easily resizable than they currently are.