Recipe Screen and Elements
The Recipe frame has now almost entirely been implemented, ready to be linked up to the program backend.  Several adjustments to the scaling are yet to be made, namely in adjusting the padding between the buttons, and removing usless/unnecessary code. Functions need to be updated later to be consistent in naming and documentation with the rest of the program.