Recipe Creation Ingredient List
The ingredient list on the recipe creation screen is now vaguely functional. It will require more tweaking to make it more intuitive to use, and throw less errors. Currently for example, the ingredient quanitity errors if a number that isn't an integer is input into it. The combo box also doesn't work, and the scaling is a bit iffy.