Recipe Submission and Deletion
Recipes can now be submitted to the main program after their creation. These recipes can also now be deleted from the system. Code was also added in to handle the creation of recipes with duplicate names, such that when they are saved they do not override the similar recipe's data