List Generation and Minor Fixes
The generate list button on the main screen is now functional, generating and saving a shopping list depending on which recipes the user has selected in the main interface. References to 'removing' have also been changed in favour of 'deletion' for consistency. A bug with loading saved recipe ingredients back into memory has also been fixed, due to the 'dual quantity type' nomenclature assigned in the program